refactor: String을 생성하는 것은 StringUtil, Int를 생성하는 것은 IntUtil
규칙을 명확히 하여 분리함
